cache: only invalidate for non-existent depends if they used to exist.
Previously, if a file listed in the cached __depends did not exist, it would
always invalidate the cache.  Now it only does so if the cached version
existed, and the current does not.  In this way, we can support storing
entries in __depends for files which may or may not exist, and ensure that
creating such a file results in cache invalidation.  Specifically, this is for
the amend.inc implementation.

Signed-off-by: Chris Larson <clarson@mvista.com>